requirements.txt
blinker==1.8.2
blowfish==0.6.1
cachelib==0.9.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
click==8.1.7
colorama==0.4.6
comtypes==1.4.8
ContFrac==1.0.0
contourpy==1.0.7
cryptography==43.0.3
cycler==0.11.0
distlib==0.3.9
dnspython==2.7.0
docx==0.2.4
docx-utils==0.1.3
docx2txt==0.8
email_validator==2.2.0
filelock==3.16.1
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Mail==0.10.0
Flask-MySQLdb==2.0.0
fonttools==4.38.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.4
lxml==4.9.2
MarkupSafe==3.0.2
matplotlib==3.7.0
mpmath==1.2.1
mysql-connector-python==9.1.0
mysqlclient==2.2.5
Naked==0.1.32
numpy==1.22.3
packaging==21.3
passlib==1.7.4
Pillow==9.1.0
pip==24.3.1
platformdirs==4.3.6
polynom==0.0.3
polynomial==0.21
pyarrow==18.0.0
pycparser==2.22
pycryptodome==3.21.0
pylance==0.19.1
pynput==1.7.6
pyOpenSSL==24.2.1
pyparsing==3.0.7
# PyQt5 ==5.15.6
# PyQt5-Qt5==5.15.2
# PyQt5-sip==12.9.1
# PyQt5-stubs==5.15.2.0
# PyQt6==6.7.0
# PyQt6-Qt6==6.7.0
# PyQt6-sip==13.6.0
PySide6==6.2.4
python-dateutil==2.8.2
python-docx==0.8.11
PyYAML==6.0.2
QtPy==2.0.1
reedsolo==1.7.0
requests==2.32.3
scipy==1.10.0
setuptools==57.0.0
shellescape==3.8.1
shiboken6==6.2.4
six==1.16.0
sympy==1.10.1
tabulate==0.9.0
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
virtualenv==20.27.1
Werkzeug==3.0.6
wheel==0.36.2
WTForms==3.2.1
XlsxWriter==3.0.3